Add files via upload
HTML, CSS, and JavaScript are core to creating web applications. In building web applications, HTML provides the basic structure of websites, which is enhanced and modified by other technologies like CSS and JavaScript. CSS is used to control presentation, formatting, styling, and layout. JavaScript is used to control the behavior of different elements in the application.